Fix the scroll issue
From Chrome 61, we can't get or set scrollTop, scrollLeft and like that
through "document.body" interface anymore because of changing its
implementation. And this changing killed the scrolling features. So I
replaced this interface to fix it.

To be precise, Chrome changed its implementation that "document.body"
was no longer a reflection of the viewport, and set
document.documentElement or document.scrollingElement as the viewport
instead, as a result of following the standardized spec.

Please see below for more details.
https://drafts.csswg.org/cssom-view/#scrolling-box
https://drafts.csswg.org/cssom-view/#dom-document-scrollingelement
https://www.w3.org/TR/DOM-Level-2-Core/core.html#ID-87CD092